requirements.txt
requests==2.31.0
pandas==2.2.1
streamlit==1.32.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
httpx==0.27.0
parsel==1.9.0
pytest==8.1.1
pytest-asyncio==0.23.5
loguru==0.7.2
plotly==5.19.0 